@@ -1,7 +1,7 @@
+import {chromium, Browser, Page, ElementHandle} from '@playwright/test'
 import NodeEnvironment from 'jest-environment-node'
 import {isEqual} from 'lodash'
 import ipc from 'node-ipc'
-import {chromium, Browser, Page, ElementHandle} from 'playwright'
 import {PortableTextBlock} from '@sanity/types'
 import {normalizeSelection} from '../../src/utils/selection'
 import type {EditorSelection} from '../../src'
