@@ -12,8 +12,20 @@ export default defineScope({
     },
     {
       name: 'form-builder',
-      title: 'Form builder',
+      title: 'Form builder (unfinished)',
       component: lazy(() => import('./FormBuilderStory')),
+    },
+    {
+      name: 'form-builder-pte-input',
+      title: 'Form builder: Portable Text Input',
+      component: lazy(
+        () => import('../../../../playwright-ct/tests/formBuilder/PortableTextInputStory'),
+      ),
+    },
+    {
+      name: 'form-builder-array-input',
+      title: 'Form builder: Array Input',
+      component: lazy(() => import('../../../../playwright-ct/tests/formBuilder/ArrayInputStory')),
     },
   ],
 })
