@@ -1,6 +1,5 @@
-import React, {useCallback, useMemo, useRef, useState} from 'react'
+import React, {useCallback, useMemo, useRef} from 'react'
 import {
-  ArraySchemaType,
   BooleanSchemaType,
   isBooleanSchemaType,
   isNumberSchemaType,
@@ -36,15 +35,7 @@ import {DEFAULT_STUDIO_CLIENT_OPTIONS} from '../../../../studioClient'
 import {readAsText} from '../../../studio/uploads/file/readAsText'
 import {accepts} from '../../../studio/uploads/accepts'
 import {applyAll} from '../../../patch/applyPatch'
-import {
-  FieldActionMenu,
-  FieldActionsProvider,
-  FieldActionsResolver,
-  useFieldActions,
-} from '../../../field'
-import {useFormPublishedId} from '../../../useFormPublishedId'
 import {useFormBuilder} from '../../../useFormBuilder'
-import {DocumentFieldActionNode} from '../../../../config'
 
 function move<T>(arr: T[], from: number, to: number): T[] {
   const copy = arr.slice()
@@ -169,8 +160,6 @@ export function ArrayOfPrimitivesField(props: {
   const {
     field: {actions: fieldActions},
   } = useFormBuilder().__internal
-  const documentId = useFormPublishedId()
-  const [fieldActionNodes, setFieldActionNodes] = useState<DocumentFieldActionNode[]>([])
 
   const focusRef = useRef<Element & {focus: () => void}>()
   const uploadSubscriptions = useRef<Subscription>()
@@ -411,10 +400,7 @@ export function ArrayOfPrimitivesField(props: {
 
   const fieldProps: Omit<ArrayOfPrimitivesFieldProps, 'renderDefault'> = useMemo(() => {
     return {
-      actions:
-        fieldActionNodes.length > 0 ? (
-          <FieldActionMenu focused={member.field.focused} nodes={fieldActionNodes} />
-        ) : undefined,
+      actions: fieldActions,
       name: member.name,
       index: member.index,
       level: member.field.level,
@@ -435,8 +421,9 @@ export function ArrayOfPrimitivesField(props: {
       inputProps: inputProps as ArrayOfPrimitivesInputProps,
     }
   }, [
-    fieldActionNodes,
-    member.field.focused,
+    fieldActions,
+    member.name,
+    member.index,
     member.field.level,
     member.field.value,
     member.field.schemaType,
@@ -445,8 +432,6 @@ export function ArrayOfPrimitivesField(props: {
     member.field.path,
     member.field.presence,
     member.field.validation,
-    member.name,
-    member.index,
     member.collapsible,
     member.collapsed,
     handleExpand,
@@ -465,20 +450,7 @@ export function ArrayOfPrimitivesField(props: {
       onPathBlur={onPathBlur}
       onPathFocus={onPathFocus}
     >
-      {documentId && fieldActions.length > 0 && (
-        <FieldActionsResolver
-          actions={fieldActions}
-          documentId={documentId}
-          documentType={member.field.schemaType.name}
-          onActions={setFieldActionNodes}
-          path={member.field.path}
-          schemaType={member.field.schemaType}
-        />
-      )}
-
-      <FieldActionsProvider actions={fieldActionNodes} path={member.field.path}>
-        {useMemo(() => renderField(fieldProps as FIXME), [fieldProps, renderField])}
-      </FieldActionsProvider>
+      {useMemo(() => renderField(fieldProps as FIXME), [fieldProps, renderField])}
     </FormCallbacksProvider>
   )
 }
